requirements.txt
discord-py
python-dotenv
asyncpraw
pysaucenao
feedparser
atproto
selenium